+++ packages/hydrogen/src/WeaverseHydrogenRoot.tsx
@@ -1,7 +1,6 @@
 import {
   isBrowser,
   type PlatformTypeEnum,
-  replaceContentDataConnectorsDeep,
   useSafeExternalStore,
   Weaverse,
   WeaverseItemStore,