+++ packages/core/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @weaverse/core
 
+## 5.5.0-beta.7
+
+### Patch Changes
+
+- update parse logic
+
 ## 5.5.0-beta.5
 
 ### Patch Changes